+++ python/lsst/ts/ess/ess_csc.py
@@ -22,6 +22,8 @@ import argparse
 import asyncio
 import json
 import platform
+from types import SimpleNamespace
+from typing import Any, Dict, List, Optional, Union
 
 import numpy as np
 
@@ -30,7 +32,13 @@ from . import __version__
 from .ess_instrument_object import EssInstrument
 from .mock.mock_temperature_sensor import MockTemperatureSensor
 from .sel_temperature_reader import SelTemperature
-from lsst.ts import salobj, tcpip
+from lsst.ts import salobj, tcpip  # type: ignore
+from lsst.ts.envsensors import (
+    DeviceConfig,
+    DeviceType,
+    Key,
+    ResponseCode,
+)
 
 """The temperature polling interval."""
 TEMPERATURE_POLLING_INTERVAL = 0.25
@@ -70,7 +78,8 @@ class EssCsc(salobj.ConfigurableCsc):
         initial_state: salobj.State = salobj.State.STANDBY,
         simulation_mode: int = 0,
     ) -> None:
-        self.config = None
+        self.config: Optional[SimpleNamespace] = None
+        self.device_configurations: Dict[str, DeviceConfig] = {}
         self._config_dir = config_dir
         super().__init__(
             name="ESS",
@@ -82,19 +91,18 @@ class EssCsc(salobj.ConfigurableCsc):
         )
 
         # Temporary mode during transition to remote mode only.
-        self.local_mode = local_mode
+        self.local_mode: bool = local_mode
 
         # Used if self.local_mode == True
-        self.device = None
-        self.ess_instrument = None
+        self.ess_instruments: List[EssInstrument] = []
 
         # Used if self.local_mode == False
-        self.reader = None
-        self.writer = None
+        self.reader: Optional[asyncio.StreamReader] = None
+        self.writer: Optional[asyncio.StreamWriter] = None
         self.host = None
         self.port = None
-        self.telemetry_loop = salobj.make_done_future()
-        self.last_commands = []
+        self.telemetry_loop: asyncio.Future = salobj.make_done_future()
+        self.last_commands: List[str] = []
 
         # Unit tests may set this to an integer value to simulate a
         # disconnected or missing sensor.
@@ -108,9 +116,8 @@ class EssCsc(salobj.ConfigurableCsc):
             while True:
                 data = await self.read()
                 if RESPONSE in data:
-                    # TODO: Properly support command responses (DM-30265)
                     response = data[RESPONSE]
-                    if response != 0:
+                    if response != ResponseCode.OK:
                         try:
                             oldest_last_command = self.last_commands.pop(0)
                             self.log.error(
@@ -121,8 +128,7 @@ class EssCsc(salobj.ConfigurableCsc):
                                 f"Received response {data} while no command was waiting for a reply."
                             )
                 if TELEMETRY in data:
-                    # TODO: Properly support multiple sensors (DM-30265)
-                    output = data[TELEMETRY][1:]
+                    output = data[TELEMETRY]
                     await self.get_telemetry(output=output)
         except Exception:
             self.log.exception("_read_loop failed")
@@ -135,21 +141,29 @@ class EssCsc(salobj.ConfigurableCsc):
         data : `dict`
             A dictionary with objects representing the string read.
         """
+        if not self.reader or not self.connected:
+            raise RuntimeError("Not connected")
+
         read_bytes = await asyncio.wait_for(
             self.reader.readuntil(tcpip.TERMINATOR), timeout=SOCKET_TIMEOUT
         )
         data = json.loads(read_bytes.decode())
         return data
 
-    async def write(self, **data: dict) -> None:
-        """Write the data appended with a newline character.
+    async def write(self, command: str, **data: Any) -> None:
+        """Write the command and data appended with a newline character.
 
         Parameters
         ----------
+        command: `str`
+            The command to write.
         data: `dict`
             The data to write.
         """
-        st = json.dumps({**data})
+        if not self.writer or not self.connected:
+            raise RuntimeError("Not connected")
+
+        st = json.dumps({"command": command, **data})
         self.last_commands.append(st)
         self.writer.write(st.encode() + tcpip.TERMINATOR)
         await self.writer.drain()
@@ -161,23 +175,35 @@ class EssCsc(salobj.ConfigurableCsc):
         ----------
         output: `list`
             An array containing the timestamp, error and temperatures as
-            measured by the sensor.
+            measured by the sensor. The order of the items in the list is:
+            - Sensor name
+            - Timestamp
+            - Response code
+            - One or more sensor data
         """
         try:
-            data = {"timestamp": output[0]}
-            error_code = output[1]
+            print(output)
+            sensor_name = output[0]
+            timestamp = output[1]
+            error_code = output[2]
+            device_configuration = self.device_configurations[sensor_name]
             if error_code == "OK":
-                for i in range(2, 2 + self.config.channels):
-                    # The telemetry channels start counting at 1 and not 0.
-                    data[f"temperatureC{i-1:02d}"] = (
-                        float.nan if np.isnan(output[i]) else output[i]
+                telemetry = {"sensor_name": sensor_name, "timestamp": timestamp}
+                sensor_data = output[3:]
+                if len(sensor_data) != device_configuration.channels:
+                    raise RuntimeError(
+                        f"Expected {device_configuration.channels} temperatures "
+                        f"but received {len(sensor_data)}."
                     )
-                self.log.info(f"Received temperatures {data}")
+                for i, value in enumerate(sensor_data):
+                    # The telemetry channels start counting at 1 and not 0.
+                    telemetry[f"temperatureC{i+1:02d}"] = value
+                self.log.info(f"Received temperatures {telemetry}")
                 self.log.info("Sending telemetry.")
                 telemetry_method = getattr(
-                    self, f"tel_temperature{self.config.channels}Ch"
+                    self, f"tel_temperature{device_configuration.channels}Ch"
                 )
-                telemetry_method.set_put(**data)
+                telemetry_method.set_put(**telemetry)
         except Exception:
             self.log.exception("Method get_telemetry() failed")
 
@@ -203,19 +229,31 @@ class EssCsc(salobj.ConfigurableCsc):
         """Connect to the ESS sensor or start the mock sensor, if in
         simulation mode.
         """
-        self.log.info("Connecting to the local sensor.")
-        self._get_device()
-        sel_temperature = SelTemperature(
-            self.config.name, self.device, self.config.channels, self.log
-        )
-        self.ess_instrument = EssInstrument(
-            self.config.name, sel_temperature, self.get_telemetry, self.log
-        )
-        self.log.info("Connection to the local sensor established.")
+        self.log.info("Connecting to the local sensor(s).")
+        for sensor_name in self.device_configurations:
+            device_configuration = self.device_configurations[sensor_name]
+            device = self._get_device(device_configuration=device_configuration)
+            sel_temperature = SelTemperature(
+                device_configuration.name,
+                device,
+                device_configuration.channels,
+                self.log,
+            )
+            self.ess_instruments.append(
+                EssInstrument(
+                    device_configuration.name,
+                    sel_temperature,
+                    self.get_telemetry,
+                    self.log,
+                )
+            )
+            self.log.info("Connection to the local sensor established.")
 
     async def connect_socket(self) -> None:
         """Connect to the SocketServer and send the configuration for which
         sensors to start reading the data of."""
+        if not self.config:
+            raise RuntimeError("Not configured yet.")
         if not self.host:
             self.host = self.config.host
         if not self.port:
@@ -224,19 +262,10 @@ class EssCsc(salobj.ConfigurableCsc):
         self.reader, self.writer = await asyncio.wait_for(
             rw_coro, timeout=SOCKET_TIMEOUT
         )
-        # TODO: Properly support multiple sensors (DM-30265)
-        configuration = {
-            "devices": [
-                {
-                    "name": self.config.name,
-                    "channels": self.config.channels,
-                    "type": self.config.type,
-                    "ftdi_id": self.config.ftdi_id,
-                }
-            ]
-        }
+        configuration = {"devices": self.config.devices}
         await self.write(
-            command="configure", parameters={"configuration": configuration}
+            command="configure",
+            parameters={"configuration": configuration},
         )
 
         # Start a loop to read incoming data from the SocketServer.
@@ -272,7 +301,8 @@ class EssCsc(salobj.ConfigurableCsc):
 
         self.log.info("Start periodic polling of the sensor data.")
         if self.local_mode:
-            await self.ess_instrument.start()
+            for ess_instrument in self.ess_instruments:
+                await ess_instrument.start()
         else:
             await self.write(command="start", parameters={})
         await super().end_enable(id_data)
@@ -291,7 +321,8 @@ class EssCsc(salobj.ConfigurableCsc):
         self.cmd_disable.ack_in_progress(id_data, timeout=60)
         if self.local_mode:
             try:
-                await self.ess_instrument.stop()
+                for ess_instrument in self.ess_instruments:
+                    await ess_instrument.stop()
             except Exception:
                 self.log.exception("Error in begin_disable. Continuing...")
         else:
@@ -307,24 +338,46 @@ class EssCsc(salobj.ConfigurableCsc):
         sensor, if running.
         """
         self.log.info("Disconnecting")
-        self.ess_instrument = None
+        self.ess_instruments = []
 
     async def configure(self, config) -> None:
         """Configure the CSC.
 
+        Also store the device configurations for easier access when receiving
+        and processing telemetry.
+
         Parameters
         ----------
         config : `object`
-            The configuration as described by the schema at ``schema_path``,
-            as a struct-like object.
+            The configuration as described by the schema at
+            `lsst.ts.ess.CONFIG_SCHEMA`, as a struct-like object.
         """
         self.config = config
+        for device in config.devices:
+            if device[Key.TYPE] == DeviceType.FTDI:
+                dev_id = Key.FTDI_ID
+            elif device[Key.TYPE] == DeviceType.SERIAL:
+                dev_id = Key.SERIAL_PORT
+            else:
+                raise ValueError(f"Unknown device type {device[Key.TYPE]} encountered.")
+            self.device_configurations[device[Key.NAME]] = DeviceConfig(
+                name=device[Key.NAME],
+                channels=device[Key.CHANNELS],
+                dev_type=device[Key.TYPE],
+                dev_id=device[dev_id],
+            )
 
     @property
     def connected(self) -> bool:
-        if self.ess_instrument is None:
-            return False
-        return True
+        if self.local_mode:
+            return self.ess_instruments is None
+        else:
+            return not (
+                self.reader is None
+                or self.writer is None
+                or self.reader.at_eof()
+                or self.writer.is_closing()
+            )
 
     @staticmethod
     def get_config_pkg() -> str:
@@ -342,44 +395,64 @@ class EssCsc(salobj.ConfigurableCsc):
         parser.add_argument(
             "--local-mode",
             default="False",
-            help="directory containing configuration files for the start command.",
+            help="Boolean indicating if local mode (True) or remote mode (False) should be used.",
             dest="local_mode",
         )
         super().add_arguments(parser)
 
-    def _get_device(self) -> None:
+    def _get_device(self, device_configuration: DeviceConfig) -> Optional[Any]:
         """Get the device to connect to by using the configuration of the CSC
         and by detecting whether the code is running on an aarch64 architecture
         or not.
 
+        Parameters
+        ----------
+        device_configuration: `dict`
+            A dict representing the device to connect to. The format of the
+            dict follows `lsst.ts.ess.CONFIG_SCHEMA`.
+
         Returns
         -------
-        device: `VcpFtdi` or `RpiSerialHat` or `None`
-
+        device: `MockTemperatureSensor` or `VcpFtdi` or `RpiSerialHat` or
+            `None`
+            The device to connect to.
 
         Raises
         ------
-
+        RuntimeError
+            In case an incorrect configuration has been loaded.
         """
-        self.device = None
+        device: Any = None
         if self.simulation_mode == 1:
             self.log.info("Connecting to the mock sensor.")
-            self.device = MockTemperatureSensor(
-                "MockSensor", 4, nan_channel=self.nan_channel
+            device = MockTemperatureSensor(
+                device_configuration.name,
+                device_configuration.channels,
+                disconnected_channel=self.nan_channel,
             )
-        elif self.config.type == "FTDI":
+        elif device_configuration.dev_type == DeviceType.FTDI:
             from .vcp_ftdi import VcpFtdi
 
-            self.device = VcpFtdi(self.config.name, self.config.ftdi_id, self.log)
-        elif self.config.type == "Serial":
+            device = VcpFtdi(
+                device_configuration.name,
+                device_configuration.dev_id,
+                self.log,
+            )
+        elif device_configuration.dev_type == DeviceType.SERIAL:
             # make sure we are on a Raspberry Pi4
             if "aarch64" in platform.platform():
                 from .rpi_serial_hat import RpiSerialHat
 
-                self.device = RpiSerialHat(self.config.name, self.config.port, self.log)
+                device = RpiSerialHat(
+                    device_configuration.name,
+                    device_configuration.dev_id,
+                    self.log,
+                )
 
-        if self.device is None:
+        if device is None:
             raise RuntimeError(
-                f"Could not get a {self.config.type!r} device on architecture "
-                f"{platform.platform()}. Please check the configuration."
+                f"Could not get a {device_configuration['type']!r} device on "
+                f"architecture {platform.platform()}. Please check the "
+                f"configuration."
             )
+        return device